search: add modal header, rename search-box
