+++ eahub/base/static/components/search-profiles/search.html
@@ -6,8 +6,12 @@
         <div class="modal fade" id="search-box-modal" aria-hidden="true" tabindex="-1">
             <div class="modal-dialog modal-fullscreen-sm-down">
                 <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title">Filters</h5>
+                        <button type="button" class="btn-close" data-dismiss="modal" aria-label="Close"></button>
+                    </div>
                     <div class="modal-body">
-                        <search-box/>
+                        <search-filters/>
                     </div>
                 </div>
             </div>
@@ -15,7 +19,7 @@
 
         <div class="row">
             <div class="d-none d-lg-flex col-lg-6 col-xl-5 col-xxl-4">
-                <search-box/>
+                <search-filters/>
             </div>
 
             <div class="col-24 col-lg-18 col-xl-19 col-xxl-20">
